[REFACTOR] domain 폴더를 entity/document로 분리
같은 domain이지만 디비와 연결하는 entity와 엘라스틱서치와 연결하는 document를 분리하는 것이 좋다고 생각
