+++ src/main/java/com/merseongsanghoe/sooljarisearchengine/document/AlcSearchKeyDocument.java
@@ -1,4 +1,4 @@
-package com.merseongsanghoe.sooljarisearchengine.domain;
+package com.merseongsanghoe.sooljarisearchengine.document;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
